+++ samples-PowerShell/Signal.far.ps1
@@ -1,0 +1,4 @@
+﻿$ErrorActionPreference=1
+Import-Module "$env:FARHOME\FarNet\Lib\FarNet.ScottPlot"
+
+Show-FarPlotSignal @(1, 4, 9, 16, 25)