+++ src/Enums.cs
@@ -1,0 +1,12 @@
+﻿namespace ScottPlot;
+
+///
+public enum FarScatterKind
+{
+    ///
+    None,
+    ///
+    Line,
+    ///
+    Points,
+}