+++ test/FSharp.test.ps1
@@ -1,6 +1,10 @@
 ﻿
 . ./About.ps1
 
+task Signal {
+	Start-Far @Testing "fs:exec file=..\samples-FSharp\Signal.fsx"
+}
+
 task Scatter {
 	Start-Far @Testing "fs:exec file=..\samples-FSharp\Scatter.fsx"
 }